* Factor out useProducts * Make ProductGrid
