+++ src/route/CmsPage/ProductGrid/grid.css
@@ -1,0 +1,12 @@
+.root {
+    display: grid;
+    grid-template-areas:
+        'actions'
+        'items';
+    grid-template-columns: 1fr;
+    line-height: 1;
+}
+
+.images {
+    
+}
