+++ src/route/CmsPage/ProductList/ProductList.js
@@ -1,65 +1,22 @@
 import React from 'react';
-import { useQuery } from 'Route/CmsPage/useQuery';
-import ProductListQuery from '@scandipwa/scandipwa/src/query/ProductList.query';
-
 import ProductCard from '@scandipwa/scandipwa/src/component/ProductCard/ProductCard.component';
+import { useProducts } from 'Route/CmsPage/useProducts';
 import './ProductList.scss';
 
 const ProductList = props => {
     const {
-        item,
-        formatMessage
-    } = props;
-    let filterData = { category_id: { eq: '6' } };
-    let sortData;
-    let pageSize = 12;
-
-    if (item.dataParsed) {
-        const { dataParsed } = item;
-        if (dataParsed.openProductsWidthSKUs) {
-            let openProductsWidthSKUs = item.dataParsed.openProductsWidthSKUs;
-            openProductsWidthSKUs = openProductsWidthSKUs.trim();
-            openProductsWidthSKUs = openProductsWidthSKUs.split(',');
-            filterData = {
-                sku: {
-                    in: openProductsWidthSKUs
-                }
-            };
-        } else if (dataParsed.openCategoryProducts) {
-            filterData = { category_id: { eq: String(dataParsed.openCategoryProducts) } };
-        }
-        if (dataParsed.openProductsWidthSortAtt) {
-            const directionToSort = dataParsed.openProductsWidthSortDir ? dataParsed.openProductsWidthSortDir.toUpperCase() : 'ASC';
-            sortData = {};
-            sortData[dataParsed.openProductsWidthSortAtt] = directionToSort;
-        }
-        if (dataParsed.openProductsWidthSortPageSize) {
-            pageSize = parseInt(dataParsed.openProductsWidthSortPageSize);
-        }
-    }
-
-    const q = ProductListQuery.getQuery({
-        args: {
-            filter: {
-                categoryIds: parseInt(filterData.category_id.eq)
-            },
-            pageSize: pageSize,
-            sort: sortData
-        }
-    });
-    const {
         data,
-        loading
-    } = useQuery(q);
+        loading,
+        canRender,
+        wholeName
+    } = useProducts(props);
 
-    if (data && data.products && data.products.items && data.products.items.length) {
-        const name = formatMessage({ val: item.name });
-        // const name = 'hello';
+    if (canRender) {
         return (
             <div className={'root product-list'}>
                 <div className={'list-title'}
                 >
-                    {name}
+                    {wholeName}
                 </div>
                 <div className={'overall-scroll'}
                 >
@@ -70,7 +27,7 @@ const ProductList = props => {
                                                 availableVisualOptions={['label']}
                                                 device={{}}
                                                 getAttribute={() => ''}
-                                                isBundleProductOutOfStock={false}
+                                                isBundleProductOutOfStock={() => false}
                                                 isConfigurableProductOutOfStock={() => false}
                                                 isPreview={true}
                                                 isWishlistEnabled={false}
